cleaned up 2 coords search
